refactor!: extract engine-core module
