+++ packages/java/engine-core/src/main/java/dev/hilla/engine/EngineConfiguration.java
@@ -1,4 +1,4 @@
-package dev.hilla.internal;
+package dev.hilla.engine;
 
 import java.io.File;
 import java.io.IOException;
