+++ packages/java/engine-core/src/main/java/dev/hilla/engine/GeneratorProcessor.java
@@ -1,4 +1,4 @@
-package dev.hilla.internal;
+package dev.hilla.engine;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
