+++ packages/java/engine-core/src/main/java/dev/hilla/engine/ParserClassPathConfiguration.java
@@ -1,4 +1,4 @@
-package dev.hilla.internal;
+package dev.hilla.engine;
 
 import java.util.Objects;
