+++ packages/java/engine-core/src/main/java/dev/hilla/engine/ParserException.java
@@ -1,4 +1,4 @@
-package dev.hilla.internal;
+package dev.hilla.engine;
 
 public final class ParserException extends RuntimeException {
     ParserException(String message) {
