+++ packages/java/engine-runtime/src/test/java/dev/hilla/internal/AbstractTaskEndpointGeneratorTest.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.util.List;
 
+import dev.hilla.engine.EngineConfiguration;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
