+++ packages/java/maven-plugin/pom.xml
@@ -108,7 +108,7 @@
     </dependency>
     <dependency>
       <groupId>dev.hilla</groupId>
-      <artifactId>engine-runtime</artifactId>
+      <artifactId>engine-core</artifactId>
       <version>${project.version}</version>
     </dependency>
   </dependencies>
